+++ src/constants/index.ts
@@ -4,3 +4,5 @@ export * from "./nav";
 export * from "./copy";
 export * from "./packages";
 export * from "./about";
+export * from "./seo";
+export * from "./assets";